Tests for strings; re-order parsers alphabetically
All data types have a (naive) parser implementation except for the
`number` type. The decision to rearrange the parsers in alphabetical
order was made to keep things orderly and tidy.

A bug was discovered when testing the `string` type caused by escape
sequences, or lack thereof. The bugs should be resolved soon as helper
parsers are introduced, separating and simplifying the logic involved.