Fixed string parser bugs and renamed input params
The string parser bugs that were breaking the test suite have been
fixed. The error was caused by the string parser incorrectly expecting
all characters between double quotes to be alphanumeric. This is not
compliant with the JSON format specification.

Some additional tests were added to prove that the parser now works as
intended.

The parameter to parser functions was previously just `i`, which is
quite ambiguous and can be difficult to spot visually. All functions now
take `input` instead of `i` to address this issue.