Parsers now take `&str` slices to handle unicode
Lucy's parser functions initially only took raw byte slices as inputs.
This is not the correct approach to take as it does not factor in the
idea of unicode code points. `char` types can be anywhere from 1-4 bytes
in length and failing to perform proper validation results in undefined
behavior.