Update README to point to elucidate
This project has been deprecated and archived. Please see elucidate, the successor of this project.

Lucy was an experiment to gauge whether it is worth building a parser and data pipeline using Rust instead of a more dynamic language with quicker development time (such as TypeScript/Node.js).

Rust and nom provide the necessary utilities to build a performant, memory-safe parser in fairly short amount of time. The benefits of Rust outweigh the cons and the decision was made to build lucy (now elucidate) in Rust.